Core: Remove `class_db.h` include from `ref_counted.h`
