+++ core/crypto/aes_context.cpp
@@ -28,7 +28,7 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#include "core/crypto/aes_context.h"
+#include "aes_context.h"
 
 #include "core/object/class_db.h"
