+++ core/crypto/aes_context.h
@@ -32,6 +32,7 @@
 
 #include "core/crypto/crypto_core.h"
 #include "core/object/ref_counted.h"
+#include "core/variant/binder_common.h"
 
 class AESContext : public RefCounted {
 	GDCLASS(AESContext, RefCounted);
