+++ core/crypto/hashing_context.h
@@ -31,6 +31,7 @@
 #pragma once
 
 #include "core/object/ref_counted.h"
+#include "core/variant/binder_common.h"
 
 class HashingContext : public RefCounted {
 	GDCLASS(HashingContext, RefCounted);
