+++ core/io/remote_filesystem_client.cpp
@@ -33,6 +33,7 @@
 #include "core/io/dir_access.h"
 #include "core/io/file_access.h"
 #include "core/io/stream_peer_tcp.h"
+#include "core/os/os.h"
 #include "core/string/string_builder.h"
 
 #define FILESYSTEM_CACHE_VERSION 1
