+++ core/string/print_string.cpp
@@ -33,6 +33,8 @@
 #include "core/core_globals.h"
 #include "core/os/os.h"
 
+#include <cstdio>
+
 static PrintHandlerList *print_handler_list = nullptr;
 static thread_local bool is_printing = false;
