+++ modules/navigation_2d/2d/nav_map_iteration_2d.h
@@ -35,6 +35,7 @@
 #include "nav_mesh_queries_2d.h"
 
 #include "core/math/math_defs.h"
+#include "core/os/rw_lock.h"
 #include "core/os/semaphore.h"
 
 class NavLinkIteration2D;
