+++ modules/navigation_2d/nav_map_2d.h
@@ -37,6 +37,7 @@
 
 #include "core/math/math_defs.h"
 #include "core/object/worker_thread_pool.h"
+#include "core/os/rw_lock.h"
 #include "servers/navigation_2d/navigation_constants_2d.h"
 
 #include <KdTree2d.h>
