+++ modules/navigation_3d/3d/nav_map_iteration_3d.h
@@ -35,6 +35,7 @@
 #include "nav_mesh_queries_3d.h"
 
 #include "core/math/math_defs.h"
+#include "core/os/rw_lock.h"
 #include "core/os/semaphore.h"
 
 class NavLinkIteration3D;
