+++ modules/navigation_3d/nav_link_3d.h
@@ -31,6 +31,7 @@
 #pragma once
 
 #include "3d/nav_base_iteration_3d.h"
+#include "core/os/rw_lock.h"
 #include "nav_base_3d.h"
 #include "nav_utils_3d.h"
