+++ modules/navigation_3d/nav_map_3d.h
@@ -37,6 +37,7 @@
 
 #include "core/math/math_defs.h"
 #include "core/object/worker_thread_pool.h"
+#include "core/os/rw_lock.h"
 #include "servers/navigation_3d/navigation_constants_3d.h"
 
 #include <KdTree2d.h>
