+++ platform/android/api/jni_singleton.h
@@ -32,7 +32,6 @@
 
 #include "java_class_wrapper.h"
 
-#include "core/config/engine.h"
 #include "core/templates/rb_map.h"
 #include "core/variant/variant.h"
 
@@ -48,34 +47,10 @@ class JNISingleton : public Object {
 	Ref<JavaObject> wrapped_object;
 
 protected:
-	static void _bind_methods() {
-		ClassDB::bind_method(D_METHOD("has_java_method", "method"), &JNISingleton::has_java_method);
-	}
+	static void _bind_methods();
 
 public:
-	virtual Variant callp(const StringName &p_method, const Variant **p_args, int p_argcount, Callable::CallError &r_error) override {
-		// Godot methods take precedence.
-		Variant ret = Object::callp(p_method, p_args, p_argcount, r_error);
-		if (r_error.error == Callable::CallError::CALL_OK) {
-			return ret;
-		}
-
-		// Check the method we're looking for is in the JNISingleton map.
-		// This is done because JNISingletons register methods differently than wrapped JavaClass / JavaObject to allow
-		// for access to private methods annotated with the @UsedByGodot annotation.
-		// In the future, we should remove access to private methods and require that JNISingletons' methods exposed to
-		// GDScript be all public, similarly to what we do for wrapped JavaClass / JavaObject methods. Doing so will
-		// also allow dropping and deprecating the @UsedByGodot annotation.
-		RBMap<StringName, MethodData>::Element *E = method_map.find(p_method);
-		if (E) {
-			if (wrapped_object.is_valid()) {
-				return wrapped_object->callp(p_method, p_args, p_argcount, r_error);
-			} else {
-				r_error.error = Callable::CallError::CALL_ERROR_INSTANCE_IS_NULL;
-			}
-		}
-		return Variant();
-	}
+	virtual Variant callp(const StringName &p_method, const Variant **p_args, int p_argcount, Callable::CallError &r_error) override;
 
 	Ref<JavaObject> get_wrapped_object() const {
 		return wrapped_object;
@@ -85,21 +60,9 @@ public:
 		return method_map.has(p_method);
 	}
 
-	void add_method(const StringName &p_name, const Vector<Variant::Type> &p_args, Variant::Type p_ret_type) {
-		MethodData md;
-		md.argtypes = p_args;
-		md.ret_type = p_ret_type;
-		method_map[p_name] = md;
-	}
+	void add_method(const StringName &p_name, const Vector<Variant::Type> &p_args, Variant::Type p_ret_type);
 
-	void add_signal(const StringName &p_name, const Vector<Variant::Type> &p_args) {
-		MethodInfo mi;
-		mi.name = p_name;
-		for (int i = 0; i < p_args.size(); i++) {
-			mi.arguments.push_back(PropertyInfo(p_args[i], "arg" + itos(i + 1)));
-		}
-		ADD_SIGNAL(mi);
-	}
+	void add_signal(const StringName &p_name, const Vector<Variant::Type> &p_args);
 
 	JNISingleton() {}
