+++ servers/camera/camera_server.h
@@ -34,6 +34,7 @@
 #include "core/object/ref_counted.h"
 #include "core/os/thread_safe.h"
 #include "core/templates/rid.h"
+#include "core/variant/binder_common.h"
 
 /**
 	The camera server is a singleton object that gives access to the various
