+++ servers/navigation_3d/navigation_path_query_result_3d.h
@@ -31,6 +31,7 @@
 #pragma once
 
 #include "core/object/ref_counted.h"
+#include "core/variant/binder_common.h"
 #include "core/variant/typed_array.h"
 #include "servers/navigation_3d/navigation_constants_3d.h"
