+++ tests/core/io/test_uds_server.cpp
@@ -38,6 +38,7 @@ TEST_FORCE_LINK(test_uds_server)
 #include "core/io/file_access.h"
 #include "core/io/stream_peer_uds.h"
 #include "core/io/uds_server.h"
+#include "core/os/os.h"
 
 #include <functional>
